Reorganizing Codebase Without Disrupting Functionality
- [+] feat(terminal/front_end.go): add support for printing prompt feedback received from the AI
- [+] fix(terminal/front_end.go): fix typo in comment
- [+] feat(terminal/front_end.go): add support for printing the number of tokens used in the AI's response
- [+] fix(terminal/front_end.go): fix typo in comment
- [+] feat(terminal/front_end.go): add support for printing a visual separator
- [+] fix(terminal/front_end.go): fix typo in comment
- [+] feat(terminal/front_end.go): add support for removing the AI prefix from the response
- [+] feat(terminal/front_end.go): add support for colorizing the response content
- [+] feat(terminal/front_end.go): add support for handling single asterisks in the response content
- [+] feat(terminal/front_end.go): add support for printing the AI's response with a typing effect
- [+] feat(terminal/front_end.go): add support for printing the footer after the AI response
- [+] fix(terminal/genai.go): remove unused import
- [+] fix(terminal/genai.go): remove unused function printResponse
- [+] fix(terminal/genai.go): remove unused function concatenateChatHistory
- [+] fix(terminal/genai.go): remove unused function handleTokenCountError
- [+] fix(terminal/genai.go): remove unused function printCurrentTokenCount
- [+] fix(terminal/genai.go): remove unused function updateAndPrintTotalTokenCount
- [+] fix(terminal/genai.go): remove unused function printVisualSeparator
- [+] fix(terminal/genai.go): remove unused function removeAIPrefix
- [+] fix(terminal/genai.go): remove unused function colorizeResponse
- [+] fix(terminal/genai.go): remove unused function handleSingleAsterisks
- [+] fix(terminal/genai.go): remove unused function printAIResponse
- [+] fix(terminal/genai.go): remove unused function printResponseFooter
